feat: create a separate component for message contents - MEED-8681 - Meeds-io/MIPs#163
